+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Honeybadger 1.7.0 (Unreleased) ##
 
+* Fail gracefully when Rack params cannot be parsed
+
+  *Joshua Wood*
+
 ## Honeybadger 1.6.1 ##
 
 * Fixes a bug in Rails 4 test task